Improve build setup with TypeScript project references.
This enables incremental TypeScript builds and only rebuild workspace
dependencies if needed.

It should "just work". When adding new workspace dependencies to
projects, the dependency has to be added to the references section in
the `tsconfig.json` as well. A script is included to automate this
process, so whenever adding new workspace packages or adding a workspace
dependency to an existing package, run `yarn update-ts-references` from
the workspace root.